requirements.txt
amqp==2.3.2
argcomplete==1.9.3
atomicwrites==1.2.1
attrs==18.2.0
awscli==1.16.34
bcrypt==3.1.4
beautifulsoup4==4.6.3
billiard==3.5.0.4
boto3==1.9.24
botocore==1.12.24
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
cfn-flip==1.0.3
chardet==3.0.4
Click==7.0
cloudinary==1.14.0
colorama==0.3.9
coverage==4.5.1
docutils==0.14
durationpy==0.5
future==0.16.0
hjson==3.0.1
hupper==1.3.1
idna==2.7
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
kombu==4.2.1
lambda-packages==0.20.0
Mako==1.0.7
MarkupSafe==1.0
mock==2.0.0
more-itertools==4.3.0
paginate==0.5.6
paginate-sqlalchemy==0.2.0
passlib==1.7.1
PasteDeploy==1.5.2
pbr==4.3.0
pkg-resources==0.0.0
placebo==0.8.2
plaster==1.0
plaster-pastedeploy==0.6
pluggy==0.7.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.7.0
pyasn1==0.4.4
pycparser==2.19
Pygments==2.2.0
PyMySQL==0.9.3
pyramid==1.9.2
pyramid-debugtoolbar==4.5
pyramid-jinja2==2.7
pyramid-mako==1.0.2
pyramid-retry==0.5
pyramid-tm==2.2
pytest==3.8.2
pytest-cov==2.6.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.5
PyYAML==3.13
redis==2.10.6
-e git+https://github.com/dejicranium/pozeet_.git@c819c7862607a8adaee310b1d65a8aecb666d2d2#egg=repoll
repoze.lru==0.7
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.12
toml==0.10.0
tqdm==4.19.1
transaction==2.2.1
translationstring==1.3
troposphere==2.3.3
Unidecode==1.0.22
urllib3==1.23
venusian==1.1.0
vine==1.1.4
waitress==1.1.0
WebHelpers==1.3
WebHelpers2==2.0
WebOb==1.8.2
WebTest==2.0.32
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
WTForms==2.1
zappa==0.47.1
zope.deprecation==4.3.0
zope.interface==4.5.0
zope.sqlalchemy==1.0
